Fix: renamed scheduleTestButton.tsx to schedule-test-button.tsx
